refactor: move legacy, units, and testing helpers into subpackages (#77)
Reorganise the legacy bridge, unit helpers, and mock-data tooling into
dedicated packages with focused modules instead of monolithic top-level
files, preserving the cellpycore.legacy and cellpycore.units import contracts.

Closes #77
