+++ .issueflows/03-solved-issues/issue77_original.md
@@ -1,0 +1,13 @@
+# Issue #77: Refactor
+
+Source: https://github.com/cellpy/cellpy-core/issues/77
+
+## Original issue text
+
+the meta data is in its own folder. That is good.
+
+legacy stuff (header_mapping, legacy, legacy_selector, maybe more, for example OldCellpyCellCore) is not. Should be.
+
+unit handling is not. Should be, I think.
+
+the module _helpers.py might also fit better somewhere else (legacy?)
