+++ HISTORY.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Reorganise legacy bridge, units, and test mock helpers into dedicated
+  subpackages (`legacy/`, `units/`, `testing/`): split implementation into
+  focused modules, update importers, and remove the old top-level modules. (#77)
 - Add pytest coverage reporting to CI: `pytest-cov` in the dev group,
   `[tool.coverage.run]` config, and `--cov=cellpycore` in the test workflow. (#69)
 - Document the per-step `<signal>_<stat>` engine contract (B1) and default unset
