+++ graphify-out/.graphify_labels.json
@@ -133,5 +133,23 @@
   "131": "Community 131",
   "132": "Community 132",
   "133": "Community 133",
-  "134": "Community 134"
+  "134": "Community 134",
+  "135": "Community 135",
+  "136": "Community 136",
+  "137": "Community 137",
+  "138": "Community 138",
+  "139": "Community 139",
+  "140": "Community 140",
+  "141": "Community 141",
+  "142": "Community 142",
+  "143": "Community 143",
+  "144": "Community 144",
+  "145": "Community 145",
+  "146": "Community 146",
+  "147": "Community 147",
+  "148": "Community 148",
+  "149": "Community 149",
+  "150": "Community 150",
+  "151": "Community 151",
+  "152": "Community 152"
 }
