+++ graphify-out/manifest.json
@@ -8,8 +8,8 @@
     "hash": "32a79032dfd2520ebc3780ce7690acc6"
   },
   "C:\\scripting\\cellpy-core\\dev\\demo_mock_data.py": {
-    "mtime": 1783076787.8079662,
-    "hash": "674daa22372582c3a3a9afbb82844033"
+    "mtime": 1783092241.4171798,
+    "hash": "797ca942364f5d059e06f80c3baf682a"
   },
   "C:\\scripting\\cellpy-core\\dev\\make_harmonized_raw.py": {
     "mtime": 1783081001.5860884,
@@ -24,52 +24,64 @@
     "hash": "381707e3244adb4962ff65e4fc4e4079"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\cell_core.py": {
-    "mtime": 1783089814.4726381,
-    "hash": "f444c5bc31418326dd7116279633d578"
+    "mtime": 1783092261.1576765,
+    "hash": "f175e25b9f96fe98af1445df7b99f2b7"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\config.py": {
-    "mtime": 1783089740.950892,
-    "hash": "0b560528e30dacd6535f0ed90c336c5c"
+    "mtime": 1783092246.7154655,
+    "hash": "d126771d6c3562a212fefdf5a2e36454"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\extractors.py": {
     "mtime": 1783076787.8558426,
     "hash": "b8a0980e0461de468accbc553c0decc5"
-  },
-  "C:\\scripting\\cellpy-core\\src\\cellpycore\\header_mapping.py": {
-    "mtime": 1783076787.8578575,
-    "hash": "9e3271cb780de400e3c01834ca015dab"
-  },
-  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy.py": {
-    "mtime": 1783089705.6977088,
-    "hash": "3610f11a48353131d6f88ac5a4887669"
-  },
-  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy_selectors.py": {
-    "mtime": 1783087630.4679413,
-    "hash": "03d672ea794e32f551ffe766bce6bdef"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\settings_base.py": {
     "mtime": 1783076787.8667142,
     "hash": "06d106e9081984b20c222fe899093b00"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\summarizers.py": {
-    "mtime": 1783089742.1420918,
+    "mtime": 1783091639.7902982,
     "hash": "8da6428e96854cb3597556e7402b4f4f"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\timestamps.py": {
     "mtime": 1783076787.8703523,
     "hash": "247aca7786ee5249aa1553e990d36071"
   },
-  "C:\\scripting\\cellpy-core\\src\\cellpycore\\units.py": {
-    "mtime": 1783089059.3409834,
-    "hash": "a396ce66e39be426f13aaa87bb8a02a8"
-  },
-  "C:\\scripting\\cellpy-core\\src\\cellpycore\\_helpers.py": {
-    "mtime": 1783076787.849579,
-    "hash": "8dae08285af5bce4e2a0f52fe84b2d45"
-  },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\__init__.py": {
-    "mtime": 1783076787.8465705,
-    "hash": "184769736397386d99a1dd637fa4ccdf"
+    "mtime": 1783092244.229037,
+    "hash": "54b1e28f62221ff83302b4e89f9cc4f3"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\exceptions.py": {
+    "mtime": 1783092394.246458,
+    "hash": "9cfbc363381062792e8ca0022d42c187"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\headers.py": {
+    "mtime": 1783092400.6980312,
+    "hash": "8943ce407fbf4fe62c87e30b9dd650f1"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\limits.py": {
+    "mtime": 1783092395.1298957,
+    "hash": "fc29a917877ff9eac555298aad20606b"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\mapping.py": {
+    "mtime": 1783076787.8578575,
+    "hash": "9e3271cb780de400e3c01834ca015dab"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\meta.py": {
+    "mtime": 1783092395.596609,
+    "hash": "be115fa4bed56371c132fdedd3a3a513"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\mock_core.py": {
+    "mtime": 1783092402.150218,
+    "hash": "c5fd99e2bb43f40d62b6a6b8530f3d6f"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\selectors.py": {
+    "mtime": 1783087630.4679413,
+    "hash": "03d672ea794e32f551ffe766bce6bdef"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\legacy\\__init__.py": {
+    "mtime": 1783092416.383329,
+    "hash": "3ba7faf24e6daaf24d57dd2928f6ae9b"
   },
   "C:\\scripting\\cellpy-core\\src\\cellpycore\\metadata\\io.py": {
     "mtime": 1782376758.101906,
@@ -83,9 +95,29 @@
     "mtime": 1783076787.860014,
     "hash": "83f89a80c4a0b35de9cf71b1e92eddcd"
   },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\testing\\mock_data.py": {
+    "mtime": 1783092217.7052577,
+    "hash": "8dae08285af5bce4e2a0f52fe84b2d45"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\testing\\__init__.py": {
+    "mtime": 1783092216.2823048,
+    "hash": "be8416d0079c5793c00927e19833a3d5"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\units\\converters.py": {
+    "mtime": 1783092391.2424786,
+    "hash": "8fed97b79c7b7190256604326ed09565"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\units\\spec.py": {
+    "mtime": 1783092384.598318,
+    "hash": "9ce1a044f4719d319336dc103b92bddf"
+  },
+  "C:\\scripting\\cellpy-core\\src\\cellpycore\\units\\__init__.py": {
+    "mtime": 1783092391.6836264,
+    "hash": "9c4ade3225811426cf90496be253f39d"
+  },
   "C:\\scripting\\cellpy-core\\tests\\conftest.py": {
-    "mtime": 1783076787.8703523,
-    "hash": "1d6843072a810a3f4531b6da6b37d8a5"
+    "mtime": 1783092260.929025,
+    "hash": "c98a122ca2e1676e54e46d43f2d0e333"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_benchmarks.py": {
     "mtime": 1783081001.6122303,
@@ -96,8 +128,8 @@
     "hash": "e7c29963cb6d9a6313aae2a9e966a686"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_creation.py": {
-    "mtime": 1783076787.881833,
-    "hash": "6d83eef013882d33583ceda29697928f"
+    "mtime": 1783092260.929025,
+    "hash": "370c961db4e2c7e0b3babe1d02ed9ddf"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_e2e.py": {
     "mtime": 1783081001.6142244,
@@ -116,16 +148,16 @@
     "hash": "ca19ca10674d3cd2d5f25c72cca60879"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_header_mapping.py": {
-    "mtime": 1783076787.8853567,
-    "hash": "bcf916d04b4971c351e570dc41759207"
+    "mtime": 1783092261.1576765,
+    "hash": "7a913d680153084517b4e882bf836b65"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_import.py": {
     "mtime": 1783076787.8897276,
     "hash": "e49bfc9e96130dd7eb71e4bb2fc35b33"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_legacy_selectors.py": {
-    "mtime": 1783087630.474258,
-    "hash": "82801f3123eb95469a3604694e7fbc56"
+    "mtime": 1783092261.1576765,
+    "hash": "0ae747f96ebca7bcdaf839eb55dc020c"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_limits.py": {
     "mtime": 1783076787.8897276,
@@ -136,8 +168,8 @@
     "hash": "99f50b43036087566a87c8de4c8e3291"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_schema.py": {
-    "mtime": 1783089771.835024,
-    "hash": "41a1189f9fe73e228b45ce4feee28c24"
+    "mtime": 1783092242.0789468,
+    "hash": "be7e2a6836eba7ec1aa2cf1126dab730"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_timestamps.py": {
     "mtime": 1782463891.2938235,
@@ -148,8 +180,8 @@
     "hash": "489c08e9a10128196cf0fde6335b7452"
   },
   "C:\\scripting\\cellpy-core\\tests\\test_units_optional.py": {
-    "mtime": 1783081001.6205494,
-    "hash": "828dfb384a6849c4941b7a373102a589"
+    "mtime": 1783092429.8238554,
+    "hash": "1ad46b364de665836481842eca7997c4"
   },
   "C:\\scripting\\cellpy-core\\tests\\__init__.py": {
     "mtime": 1783076787.8703523,
@@ -160,8 +192,8 @@
     "hash": "7a39a20dddacd1c2c4f9a24b8b625907"
   },
   "C:\\scripting\\cellpy-core\\HISTORY.md": {
-    "mtime": 1783089059.3341591,
-    "hash": "7bddb39f8a9c4b8841ebdb12f2a90e15"
+    "mtime": 1783091817.5799646,
+    "hash": "c07e27daa3c3fa23d7d240cfda09e3e0"
   },
   "C:\\scripting\\cellpy-core\\README.md": {
     "mtime": 1783076787.8010132,
@@ -184,7 +216,7 @@
     "hash": "73315510f482a4fdc158dd523a4570e3"
   },
   "C:\\scripting\\cellpy-core\\docs\\standalone-use.md": {
-    "mtime": 1783089747.8341055,
+    "mtime": 1783091639.758927,
     "hash": "339f27363d431cd56dec14d52f4dd1e4"
   },
   "C:\\scripting\\cellpy-core\\docs\\data_format_specifications\\cycle_table.md": {
