+++ src/cellpycore/legacy/headers.py
@@ -1,95 +1,11 @@
-# contains mocks and legacy code to help with the migration to cellpy core
+"""Legacy cellpy column-header dataclasses (``Headers*``)."""
 
-import logging
-import numbers
 from dataclasses import dataclass
-from typing import List, Optional
+from typing import List
 
-from cellpycore.config import STEP_TYPES  # noqa: F401  (re-exported; see note below)
-
-# ``DictLikeClass``/``BaseSettings`` now live in ``cellpycore.settings_base`` and
-# ``CellpyUnits`` in ``cellpycore.units`` (STEP-12 promotion, issue #40). They are
-# re-imported here so existing ``cellpycore.legacy`` imports (and the cellpy<->core
-# settings-parity contract test) keep resolving them from this module.
-from cellpycore.settings_base import BaseSettings, DictLikeClass  # noqa: F401
-from cellpycore.units import CellpyUnits  # noqa: F401
-
-logger = logging.getLogger(__name__)
+from cellpycore.settings_base import BaseSettings
 
 
-class CellpyError(Exception):
-    """Base class for other exceptions"""
-
-    pass
-
-
-class NoDataFound(CellpyError):
-    """Exception raised when no data is found"""
-
-    pass
-
-
-# Canonical list of step-type labels produced by step-type detection in
-# ``summarizers.make_step_table`` (mirrors cellpy). The vocabulary is now
-# defined once as ``config.StepType``; ``STEP_TYPES`` is imported from there
-# (above) so this module keeps its historical name. ``CAPACITY_MODIFIERS`` lists
-# step labels that modify the running capacity (reserved for future use).
-CAPACITY_MODIFIERS = ["reset"]
-
-
-# -----------------------------------------------------------------------------
-#   Old cellpy limits class (step-type detection thresholds).
-# -----------------------------------------------------------------------------
-@dataclass
-class CellpyLimits(BaseSettings):
-    """Thresholds used when classifying step types in ``make_step_table``.
-
-    Since all instruments have an inherent inaccuracy, it is naive to assume
-    that e.g. the voltage within a constant-voltage step does not change at all.
-    These limits define what counts as "constant" and what is assumed to be
-    zero.
-
-    Verbatim mirror of ``cellpy.parameters.internal_settings.CellpyLimits``
-    (field names and default values), kept here so cellpy-core can build a step
-    table standalone and so the bridge copies stay contract-comparable (see
-    jepegit/cellpy#378). The keys match :data:`summarizers.DEFAULT_RAW_LIMITS`.
-    """
-
-    current_hard: float = 1e-13
-    current_soft: float = 1e-05
-    stable_current_hard: float = 2.0
-    stable_current_soft: float = 4.0
-    stable_voltage_hard: float = 2.0
-    stable_voltage_soft: float = 4.0
-    stable_charge_hard: float = 0.9
-    stable_charge_soft: float = 5.0
-    ir_change: float = 1e-05
-
-
-# -----------------------------------------------------------------------------
-#   Old cellpy meta class.
-# -----------------------------------------------------------------------------
-@dataclass
-class Meta:
-    pass
-
-
-class MockMetaTestDependent(Meta):
-    """Graceful-degradation metadata placeholder on ``Data``.
-
-    ``cycle_mode`` is optional. ``None`` (the default) means normal full-cell /
-    cathode convention at the engine boundary; set ``"anode"`` explicitly for
-    half-cell (inverted) tests. Legacy cellpy historically defaulted to
-    ``"anode"`` — the cellpy bridge must set that when loading real cells, not
-    rely on this placeholder default.
-    """
-
-    cycle_mode: Optional[str] = None
-
-
-# -----------------------------------------------------------------------------
-#   Old cellpy headers class.
-# -----------------------------------------------------------------------------
 @dataclass
 class BaseHeaders(BaseSettings):
     """Subclass of BaseSetting including option to add postfixes.
@@ -282,63 +198,3 @@ class HeadersStepTable(BaseHeaders):
     internal_resistance: str = "ir"
     internal_resistance_change: str = "ir_pct_change"
     rate_avr: str = "rate_avr"
-
-
-class MockCore:
-    def __init__(self):
-        self.cellpy_units = CellpyUnits()
-
-    # Candidates for cellpy core extension.
-    def _dump_cellpy_unit(self, value, parameter):
-        """Parse for unit, update cellpy_units class, and return magnitude."""
-        import numpy as np
-
-        c_value, c_unit = self._check_value_unit(value, parameter)
-        if not isinstance(c_value, numbers.Number) or np.isnan(c_value):
-            logger.critical(f"Could not parse {parameter} ({value})")
-            logger.critical("Setting it to 1.0")
-            return 1.0
-        if c_unit is not None:
-            self.cellpy_units[parameter] = f"{c_unit}"
-            logger.debug(f"Updated your cellpy_units['{parameter}'] to '{c_unit}'")
-
-        return c_value
-
-    @staticmethod
-    def _check_value_unit(value, parameter) -> tuple:
-        """Check if value is a valid number, or a quantity with units."""
-        if isinstance(value, numbers.Number):
-            return value, None
-        logger.critical(f"Parsing {parameter} ({value})")
-
-        try:
-            from cellpycore.units import Q
-
-            c = Q(value)
-            c_unit = c.units
-            c_value = c.magnitude
-        except ValueError:
-            logger.debug(f"Could not parse {value}")
-            return None, None
-        return c_value, c_unit
-
-
-# NOT USED
-def set_col_first(df, col_names):
-    """Set selected columns first in a pandas.DataFrame.
-
-    This function sets cols with names given in  col_names (a list) first in
-    the DataFrame. The last col in col_name will come first (processed last)
-
-    """
-
-    column_headings = df.columns
-    column_headings = column_headings.tolist()
-    try:
-        for col_name in col_names:
-            column_headings.pop(column_headings.index(col_name))
-            column_headings.insert(0, col_name)
-
-    finally:
-        df = df.reindex(columns=column_headings)
-        return df
