+++ src/cellpycore/legacy/selectors.py
