+++ src/cellpycore/testing/__init__.py
@@ -1,0 +1,4 @@
+"""Test and development helpers (mock data, fixtures support).
+
+Not part of the stable public API.
+"""
