+++ src/cellpycore/units/__init__.py
@@ -1,0 +1,19 @@
+"""Unit-spec and optional pint-backed conversion helpers for cellpy-core."""
+
+from cellpycore.units.converters import (
+    Q,
+    get_cellpy_units,
+    get_converter_to_specific,
+    get_default_output_units,
+    nominal_capacity_as_absolute,
+)
+from cellpycore.units.spec import CellpyUnits
+
+__all__ = [
+    "CellpyUnits",
+    "Q",
+    "get_cellpy_units",
+    "get_converter_to_specific",
+    "get_default_output_units",
+    "nominal_capacity_as_absolute",
+]
