+++ tests/test_legacy_selectors.py
@@ -5,10 +5,9 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from cellpycore import legacy_selectors
 from cellpycore.cell_core import Data, OldCellpyCellCore
 from cellpycore.config import legacy_schema
-from cellpycore.legacy import HeadersNormal, HeadersStepTable
+from cellpycore.legacy import HeadersNormal, HeadersStepTable, selectors
 
 DATA_DIR = Path(__file__).parent / "data"
 CYCLER_CC_RAW = DATA_DIR / "cycler_cc_raw.parquet"
@@ -43,15 +42,15 @@ def test_legacy_schema_default_avoids_attribute_error():
     """``schema=None`` must use legacy headers, not native ``default_schema()``."""
     data, steptable = _minimal_legacy_data()
 
-    cycles = legacy_selectors.get_cycle_numbers(data, steptable=steptable)
+    cycles = selectors.get_cycle_numbers(data, steptable=steptable)
     assert set(cycles) == {1, 2}
 
-    steps = legacy_selectors.get_step_numbers(
+    steps = selectors.get_step_numbers(
         data, steptype="charge", cycle_number=1, steptable=steptable
     )
     assert steps == {1: [1]}
 
-    rates = legacy_selectors.get_rates(data, steptable=steptable)
+    rates = selectors.get_rates(data, steptable=steptable)
     schema = legacy_schema()
     assert set(rates.columns) == {
         schema.step.cycle,
@@ -62,16 +61,14 @@ def test_legacy_schema_default_avoids_attribute_error():
 
 def test_get_cycle_numbers_from_raw_when_steptable_none():
     data, _ = _minimal_legacy_data()
-    cycles = legacy_selectors.get_cycle_numbers(data)
+    cycles = selectors.get_cycle_numbers(data)
     assert set(cycles) == {1, 2}
 
 
 def test_no_removed_create_selector_functions():
     """The pandas selector pair was removed once cellpy migrated off it (#45)."""
     for name in ("create_selector", "summary_selector_exluder"):
-        assert not hasattr(legacy_selectors, name), (
-            f"legacy_selectors.{name} should not exist"
-        )
+        assert not hasattr(selectors, name), f"selectors.{name} should not exist"
 
 
 @pytestmark_golden
@@ -82,15 +79,13 @@ def test_legacy_selectors_on_golden_step_table():
     data.raw = pd.read_parquet(CYCLER_CC_RAW)
     core.make_core_step_table(data, nom_cap=1.0)
 
-    cycles = legacy_selectors.get_cycle_numbers(data)
+    cycles = selectors.get_cycle_numbers(data)
     assert len(cycles) == CYCLER_CC_N_CYCLES
 
-    charge_steps = legacy_selectors.get_step_numbers(
-        data, steptype="charge", cycle_number=1
-    )
+    charge_steps = selectors.get_step_numbers(data, steptype="charge", cycle_number=1)
     assert 1 in charge_steps
     assert charge_steps[1] != [0]
 
-    rates = legacy_selectors.get_rates(data)
+    rates = selectors.get_rates(data)
     schema = legacy_schema()
     assert schema.step.rate_avr in rates.columns
